refactor(bill): use model utils for formatting and update UI
Updated the bill feature to use model utils for formatting and updated UI components.

- Replaced `core.ui.utils` with `core.model.utils` for formatting dates, numbers, and big decimals.
- Updated `BillDetailsTap` to use new model formatting utils.
- Updated `AddBillPaymentScreen` to use new model formatting utils.
- Updated `BillPaymentTap` to use new model formatting utils.
- Updated `CreationBillContent` to use new model formatting utils.
- Updated `BillScreen` to use new model formatting utils.
- Updated `BillHistoryTap` to use new model formatting utils.
- Updated `BillItemFormScreen` to use new model formatting utils.
- Updated `BillsContent` to use new model formatting utils.
- Updated `BillViewModel` to use `stateInWhileSubscribed` instead of `stateIn`.
- Updated imports and removed unused ones.
