+++ feature/bill/src/main/kotlin/com/casecode/pos/feature/bill/BillsContent.kt
@@ -60,7 +60,6 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import com.casecode.pos.core.data.utils.toFormattedDateString
 import com.casecode.pos.core.designsystem.component.PosBackground
 import com.casecode.pos.core.designsystem.component.PosEmptyScreen
 import com.casecode.pos.core.designsystem.component.PosFilterChip
@@ -76,9 +75,10 @@ import com.casecode.pos.core.designsystem.icon.PosIcons
 import com.casecode.pos.core.designsystem.theme.POSTheme
 import com.casecode.pos.core.model.data.users.PaymentStatus
 import com.casecode.pos.core.model.data.users.SupplierInvoice
+import com.casecode.pos.core.model.utils.toBigDecimalFormatted
+import com.casecode.pos.core.model.utils.toFormattedDateString
 import com.casecode.pos.core.ui.FilterSharedElementKey
 import com.casecode.pos.core.ui.parameterprovider.SupplierInvoiceParameterProvider
-import com.casecode.pos.core.ui.utils.toBigDecimalFormatted
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
