+++ feature/bill/src/main/kotlin/com/casecode/pos/feature/bill/BillsScreen.kt
@@ -158,7 +158,6 @@ fun BillsScreen(
                                 contentDesc = "LoadingBills",
                             )
                         }
-
                         is BillsUiState.Success -> {
                             BillsContent(
                                 filterScreenVisible = filtersVisible,
