+++ feature/bill/src/main/kotlin/com/casecode/pos/feature/bill/creation/BillItemFormScreen.kt
@@ -54,9 +54,9 @@ import com.casecode.pos.core.designsystem.component.PosTextButton
 import com.casecode.pos.core.designsystem.component.PosTopAppBar
 import com.casecode.pos.core.designsystem.icon.PosIcons
 import com.casecode.pos.core.model.data.users.Item
+import com.casecode.pos.core.model.utils.toFormattedString
 import com.casecode.pos.core.ui.TrackScreenViewEvent
 import com.casecode.pos.core.ui.utils.MAX_CURRENCY_LENGTH_SIZE
-import com.casecode.pos.core.ui.utils.toFormattedString
 import com.casecode.pos.feature.bill.R
 import timber.log.Timber
 import com.casecode.pos.core.ui.R.string as uiString
