+++ feature/bill/src/main/kotlin/com/casecode/pos/feature/bill/detials/BillHistoryTap.kt
@@ -39,12 +39,12 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import com.casecode.pos.core.data.utils.toFormattedDateTimeString
 import com.casecode.pos.core.designsystem.component.PosBackground
 import com.casecode.pos.core.designsystem.theme.POSTheme
 import com.casecode.pos.core.model.data.users.PaymentDetails
 import com.casecode.pos.core.model.data.users.PaymentMethod
 import com.casecode.pos.core.model.data.users.SupplierInvoice
+import com.casecode.pos.core.model.utils.toFormattedDateTimeString
 import com.casecode.pos.core.ui.parameterprovider.SupplierInvoiceParameterProvider
 import com.casecode.pos.feature.bill.R
 import kotlinx.datetime.Clock.System
