+++ feature/bill/src/main/kotlin/com/casecode/pos/feature/bill/detials/BillPaymentTap.kt
@@ -35,7 +35,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import com.casecode.pos.core.data.utils.toFormattedDateString
 import com.casecode.pos.core.designsystem.component.PosBackground
 import com.casecode.pos.core.designsystem.component.PosEmptyScreen
 import com.casecode.pos.core.designsystem.icon.PosIcons
@@ -43,9 +42,10 @@ import com.casecode.pos.core.designsystem.theme.POSTheme
 import com.casecode.pos.core.model.data.users.PaymentDetails
 import com.casecode.pos.core.model.data.users.PaymentMethod
 import com.casecode.pos.core.model.data.users.SupplierInvoice
+import com.casecode.pos.core.model.utils.toBigDecimalFormatted
+import com.casecode.pos.core.model.utils.toFormattedDateString
 import com.casecode.pos.core.ui.DevicePreviews
 import com.casecode.pos.core.ui.parameterprovider.SupplierInvoiceParameterProvider
-import com.casecode.pos.core.ui.utils.toBigDecimalFormatted
 import com.casecode.pos.feature.bill.R
 import com.casecode.pos.feature.bill.toPaymentMethodRes
 import kotlinx.datetime.Clock
