+++ feature/bill/src/main/kotlin/com/casecode/pos/feature/bill/detials/UpdateBillScreen.kt
@@ -55,7 +55,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.casecode.pos.core.data.utils.toFormattedDateString
 import com.casecode.pos.core.designsystem.component.PosDatePickerDialog
 import com.casecode.pos.core.designsystem.component.PosElevatedCard
 import com.casecode.pos.core.designsystem.component.PosEmptyScreen
@@ -68,6 +67,7 @@ import com.casecode.pos.core.designsystem.icon.PosIcons
 import com.casecode.pos.core.designsystem.theme.POSTheme
 import com.casecode.pos.core.model.data.users.Item
 import com.casecode.pos.core.model.data.users.SupplierInvoice
+import com.casecode.pos.core.model.utils.toFormattedDateString
 import com.casecode.pos.core.ui.TrackScreenViewEvent
 import com.casecode.pos.feature.bill.R
 import com.casecode.pos.feature.bill.creation.BillInputState
